Correción

+++ app/controllers/usuarios/delete_usuarios.php
@@ -1,17 +1,29 @@
 <?php
 
 include('../../config.php');
-    
-    $id_usuario = $_POST['id_usuario'];
 
-    // Preparar la sentencia con marcadores de posición
-    $sentencia = $conn->prepare("DELETE FROM usuarios WHERE id_usuario = :id_usuario");
-    // Vincular los valores a los marcadores de posición
-    $sentencia->bindParam(':id_usuario', $id_usuario, PDO::PARAM_STR);
-    // Ejecutar la sentencia
-    $sentencia->execute();
+$id_usuario = $_POST['id_usuario'];
+
+// Comprobar que el ID de usuario existe
+$sql = "SELECT * FROM usuarios WHERE id_usuario = :id_usuario";
+$query = $conn->prepare($sql);
+$query->bindParam(':id_usuario', $id_usuario, PDO::PARAM_STR);
+$query->execute();
+
+if ($query->rowCount() == 0) {
     session_start();
-    $_SESSION['mensaje_exito']="Usuario Eliminado";
-    header('Location:'.$URL.'/usuarios/');
+    $_SESSION['mensaje_error'] = 'El ID de usuario no existe';
+    header('Location: ' . $URL . '/usuarios/');
+    exit;
+}
+
+// Eliminar el usuario
+$sentencia = $conn->prepare("DELETE FROM usuarios WHERE id_usuario = :id_usuario");
+$sentencia->bindParam(':id_usuario', $id_usuario, PDO::PARAM_STR);
+$sentencia->execute();
+
+session_start();
+$_SESSION['mensaje_exito'] = 'Usuario eliminado';
+header('Location:' . $URL . '/usuarios/');
 
 ?>
